Refactor and Make Similar/NowPlayingMoviesDataSourcePrefetching into one class.
